+++ callng/utilities.lua
@@ -13,13 +13,15 @@ syslog = require("posix.syslog")
 json = require("json")
 redis = require("redis")
 require("callng.configuration")
-nglog = require("callng.nglog")
-nglog.stacks = {
+log = require("callng.nglog")
+log.stacks = {
     console =   LOGSTACK_CONSOLE,
     file    =   LOGSTACK_FILE,
     syslog  =   LOGSTACK_SYSLOG
 }
-nglog.level, nglog.host, nglog.name = LOGLEVEL, NODEID, 'libresbc'
+log.level   = LOGLEVEL
+log.host    = NODEID
+log.name    = LIBRESBC
 
 random = math.random
 unpack = _G.unpack or table.unpack
@@ -155,21 +157,17 @@ end
 
 ----------------------------------------------------------------------------
 --- switch log
-log = {
+swlog = {
     debug    = function(text, ...) freeswitch.consoleLog("debug",   string.format(text.."\n", ...)) end,
     info     = function(text, ...) freeswitch.consoleLog("info",    string.format(text.."\n", ...)) end,
     notice   = function(text, ...) freeswitch.consoleLog("notice",  string.format(text.."\n", ...)) end,
     warning  = function(text, ...) freeswitch.consoleLog("warning", string.format(text.."\n", ...)) end,
     error    = function(text, ...) freeswitch.consoleLog("err",     string.format(text.."\n", ...)) end,
     critical = function(text, ...) freeswitch.consoleLog("crit",    string.format(text.."\n", ...)) end,
-    alert    = function(text, ...) freeswitch.consoleLog("alert",   string.format(text.."\n", ...)) end
+    alert    = function(text, ...) freeswitch.consoleLog("alert",   string.format(text.."\n", ...)) end,
+    emerg    = function(text, ...) freeswitch.consoleLog("alert",   string.format(text.."\n", ...)) end
 }
-
--- back to nglog
-if LOGSTACK_CONSOLE or LOGSTACK_FILE or LOGSTACK_SYSLOG then
-    log = nglog
-end
-
+log.emlt = swlog
 ----------------------------------------------------------------------------
 
 -- CURL REQUEST